Consolidate compression code into a single header
