+++ src/parquet/CMakeLists.txt
@@ -17,10 +17,12 @@
 
 # Headers: top level
 install(FILES
+  compression.h
   exception.h
   types.h
   DESTINATION include/parquet)
 
+ADD_PARQUET_TEST(compression-test)
 ADD_PARQUET_TEST(public-api-test)
 ADD_PARQUET_TEST(types-test)
 ADD_PARQUET_TEST(reader-test)
