+++ src/parquet/file/file-deserialize-test.cc
@@ -27,7 +27,7 @@
 #include <vector>
 
 #include "parquet/column/page.h"
-#include "parquet/compression/codec.h"
+#include "parquet/compression.h"
 #include "parquet/exception.h"
 #include "parquet/file/reader-internal.h"
 #include "parquet/thrift/parquet_types.h"
