+++ src/parquet/file/metadata.h
@@ -24,7 +24,7 @@
 
 #include "parquet/column/properties.h"
 #include "parquet/column/statistics.h"
-#include "parquet/compression/codec.h"
+#include "parquet/compression.h"
 #include "parquet/schema/descriptor.h"
 #include "parquet/types.h"
 #include "parquet/util/memory.h"
